Version 0.1, working on two Raspberry Pi's
This version allows two peers on the same wifi network to handshake each other
and create a virtual ethernet interface for that peer using a static IP. However
the client generates the ip incrementally and does not query one from a DHCP server.

This version also has DNS built into it but is untested. The raspberry pi's are connected
via an ad-hoc wifi network (using the wlan0 interface) and are connected through ethernet
to two seperate networks. Pinging the virutal IP of a peer will cause the ping to be forwared
to that peer. Right now the virtual interface generated is never removed